requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
frozenlist==1.3.3
idna==3.4
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
multidict==6.0.3
mysql-connector-python==8.0.31
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
pyparsing==3.0.9
pytest==7.2.0
pytest-aiohttp==1.0.4
pytest-asyncio==0.20.3
python-dotenv==0.21.0
python-json-logger==2.0.4
requests==2.28.1
requests-mock==1.10.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
